Send Just The Warehouse Sku to Quiet
No longer worry about sets or skus being incorrectly formatted. The
variant will have knowledge of the most recent warehouse sku and we
should send this down, or up to Quiet. Main PR linked below. This will
need to be merged into the support-physical-card branch when that PR is
closed.

GitHub:
https://github.com/glossier/glossier-v2/pull/1750