Provide rubocop for linting
This commit is a squashed commit for all the linting commits made. We're
now getting a thumbsup from Rubocop but we have work to do still since
most Metrics cops have been disabled. Below are the individual commit
messages.

- RubyMine in particular but other editors too will want rubocop to be
bundled with the project or at the very least installed and this commit
takes care of that via bundler.

- Provide rubocop configs and set a default rake task

- Wholesale disabling of complexity metrics

We'll come back to this but there's no universe where we can make an
impact on this in a timely manner. Leftover warnings and critical
warnings are in the 20s and merit manual intervention.

- Delinting of hushed/inventory_unit

I am going with Rubocop's opinion about extending Structs. It's an
opinion that I share structs are a data container not meant for
inheritance.

According to the file name and it is confirmed by looking around the
project we're only surfacing InventoryUnit and Variant and Product were
meant to be private, otherwise the file name would have been models.rb.
I'm giving private_constant a shot here.

Resources about Structs:
https://thepugautomatic.com/2013/08/struct-inheritance-is-overused/
http://www.rubydoc.info/gems/rubocop/0.29.1/RuboCop/Cop/Style/StructInheritance

- Delinting of extend self

I am going with Rubocop recommendation to switch to module_function.

Resource on extend self:
http://idiosyncratic-ruby.com/8-self-improvement.html
http://www.rubydoc.info/github/bbatsov/rubocop/Rubocop/Cop/Style/ModuleFunction

- Delinting of predicate methods

- Turn off cyclomatic and abc metrics

- Delinting of hushed/blackboard

- Delinting of the double negation (!!)

Reference on double negation:
http://www.rubydoc.info/github/bbatsov/rubocop/Rubocop/Cop/Style/DoubleNegation

- Delinting of unused variables

- Delinting of variable shadowing

- Delinting of error suppression