Remove all warnings from test output (+ color)
This change needed to happen the testing output was both hard to parse
and overly verbose; I really couldn't care less that the aws sdk has an
unused variable here and there (there was like 100s of these).

We now use Rubocop on the source of this library and this feature was
redundant anyway.

Dev goals: no warnings or output of any kind in my tests only dots and
Fs and Es