Remove the remote specs
This is an interesting problem, on the one end it's nice to know that we
can put things in the S3 bucket and it's also nice to know that we can
create SQS messages but this means we have to use our real QL test
credentials in this lib .env and on Travis. I fully trust that given the
right set of configs the implementation does what it is supposed to so I
am opting out of those remote tests.

The host application should be responsible for proving through tests it has
the right credentials.