Patch a timezone issue we possibly have
The code was meant to run for UTC-5 and Travis is kindly reminding us
all that timezones are tricky by showing us the failure. I hope this is
the fix but we will see. Another possible approach would be to pull in
ActiveSupport::TimeZone support but I feel that's overkill.